put countdown logic back in rules
